refactoring gridfs connections to controller
